+++ src/simlar/indexes/__init__.py
@@ -1,7 +1,15 @@
+from .bm25x_index import BM25xIndex
 from .helix_index import HelixIndex
 from .lookup_index import LookupIndex
 from .relevance_index import RelevanceIndex
 from .simlar_engine import SimlarEngine
 from .streaming_index import StreamingHelixIndex
 
-__all__ = ["LookupIndex", "RelevanceIndex", "SimlarEngine", "StreamingHelixIndex", "HelixIndex"]
+__all__ = [
+    "BM25xIndex",
+    "LookupIndex",
+    "RelevanceIndex",
+    "SimlarEngine",
+    "StreamingHelixIndex",
+    "HelixIndex",
+]